% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ind_col_scale.R
\name{add_ind_col_scale}
\alias{add_ind_col_scale}
\title{Add individualism and collectivism aggregated scores}
\usage{
add_ind_col_scale(df)
}
\arguments{
\item{df}{a dataframe that contains the raw scales}
}
\value{
The function returns the dataframe with the
additional aggregated scores added in four new
variables rowwise: `hor_ind`, `ver_ind`, `hor_col`,
`ver_col`.
}
\description{
The function takes a dataframe with the raw individual and
collectivism scales and calculates the aggreageted scores rowwise.
}
