% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_inferences.R
\name{aggregate_inferences}
\alias{aggregate_inferences}
\title{Create summary table}
\usage{
aggregate_inferences(combinations, out_list)
}
\arguments{
\item{combinations}{Tamas please add this}

\item{out_list}{Tamas please add this}
}
\value{
The function returns a tibble that shows correct
and incorrect inference rates for different scenarios
}
\description{
Summarize the result of several simulations.
}
