% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_additional_table.R
\name{calculate_additional_table}
\alias{calculate_additional_table}
\title{Calculate the table stats for the additional analysis}
\usage{
calculate_additional_table(data)
}
\arguments{
\item{data}{the piped reformatted data for the additional anaylsis}
}
\description{
This is a super lazy function that needs to be generalized.
}
