% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contact_stat.R
\name{calculate_contact_stat}
\alias{calculate_contact_stat}
\title{Calculates the statistics for the main effect of physical contact}
\usage{
calculate_contact_stat(data = NULL, vars = NULL, label = NULL, rscale)
}
\arguments{
\item{data}{data frame or tibble that contains Region and specified variables}

\item{vars}{the variables to be collapsed}

\item{label}{what data label should be attached to the output}

\item{rscale}{prior to be passed to the ttestBF function}
}
\value{
a tibble that contains the selected statistics
  in a format that can be printed
}
\description{
This is the description
}
