% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_culture_utilitarianism.R
\name{calculate_culture_utilitarianism}
\alias{calculate_culture_utilitarianism}
\title{Effect of cultural variables on overall utilitarian responding}
\usage{
calculate_culture_utilitarianism(df = NULL, study_type)
}
\arguments{
\item{df}{a dataframe, either "study2a" or "study2b"}

\item{study_type}{the type of the study either "trolley" or "speedboat"}
}
\value{
The function returns the Bayesian and frequentist
  stats for study2 3rd order interactions.
}
\description{
Effect of cultural variables on overall utilitarian responding
}
