% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_study2_stat.R
\name{calculate_study2_stat}
\alias{calculate_study2_stat}
\title{Runs the statistical analysis for study2}
\usage{
calculate_study2_stat(data = NULL, vars = NULL, label = NULL, rscaleFixed)
}
\arguments{
\item{data}{data frame or tibble that contains Region and specified variables}

\item{vars}{the variables to be collapsed}

\item{label}{what data label should be attached to the output}

\item{rscaleFixed}{prior}
}
\value{
a tibble that contains the selected statistics
  in a format that can be printed
}
\description{
The function runs the statistical analysis for study2 a and b as
described in the manuscript.
}
