% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_country.R
\name{create_plot_country}
\alias{create_plot_country}
\title{Create country and effect size scatter plot}
\usage{
create_plot_country(data)
}
\arguments{
\item{data}{the prepared datatable for plotting}
}
\value{
The function returns a ggplot2 list object.
}
\description{
This function creates a ....
}
