% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_country_utilitarianism.R
\name{create_plot_country_utilitarianism}
\alias{create_plot_country_utilitarianism}
\title{Create country and average moral acceptability rating scatter plot}
\usage{
create_plot_country_utilitarianism(data)
}
\arguments{
\item{data}{the prepared datatable for plotting}
}
\value{
The function returns a ggplot2 list object.
}
\description{
This function creates a ....
}
