% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_ind_col.R
\name{create_plot_ind_col}
\alias{create_plot_ind_col}
\title{Create a plot to visualize the effect of individualism/collectivism scale on moral dilemmas}
\usage{
create_plot_ind_col(data)
}
\arguments{
\item{data}{datatable to be plotted}
}
\value{
The function returns a ggplot2 list object.
}
\description{
This function ...
}
