% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_study2.R
\name{create_plot_study2}
\alias{create_plot_study2}
\title{Create plot for study2}
\usage{
create_plot_study2(data)
}
\arguments{
\item{data}{the datatable to be plotted}
}
\value{
The function returns a ggplot2 list object.
}
\description{
The function creates plot for the
trolley problem paper specifically from already prepared
data.
}
