% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_values.R
\name{draw_values}
\alias{draw_values}
\title{Generate random values based on provided parameters}
\usage{
draw_values(n, mean, sd, floor = 1, ceiling = 9)
}
\arguments{
\item{n}{Number of samples}

\item{mean}{Population mean}

\item{sd}{Population SD}

\item{floor}{Minimum value}

\item{ceiling}{Maximum value}
}
\value{
A vector of n random values.
}
\description{
This function samples n integer values between a floor and a ceiling.
The probability of the values are calculated based on a
truncated normal distribution of a known mean and standard deviation.
}
