% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a dataset of integer values with specified distribution paramaters}
\usage{
generate_dataset(
  n,
  mean,
  sd,
  floor = 1,
  ceiling = 9,
  tasks = 4,
  groups = c("a", "b", "c")
)
}
\arguments{
\item{n}{TODO}

\item{mean}{TODO}

\item{sd}{TODO}

\item{floor}{TODO}

\item{ceiling}{TODO}

\item{tasks}{TODO}

\item{groups}{TODO}
}
\value{
A tibble in long format.
}
\description{
Please add a description Tamas.
}
\examples{
\dontrun{
generate_dataset(n = 1000, mean = 5, sd = 2, tasks = 4, groups = c("a", "b"))
}
}
