% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmbf_inference.R
\name{get_lmbf_inference}
\alias{get_lmbf_inference}
\title{Evaluate results lm}
\usage{
get_lmbf_inference(
  df,
  groups = c("study", "class"),
  prior = prior,
  bf_thresholds = c(10, 1/10)
)
}
\arguments{
\item{df}{TODO}

\item{groups}{TODO}

\item{prior}{TODO}

\item{bf_thresholds}{TODO}
}
\value{
A tibble with the grouping variables, a calculated BF,
and the inference.
}
\description{
Run a Bayesian ANOVA and return the BF.
}
