% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ttestbf_inference.R
\name{get_ttestbf_inference}
\alias{get_ttestbf_inference}
\title{Evaluate results ttest}
\usage{
get_ttestbf_inference(
  df,
  groups = c("study", "class"),
  formula = NULL,
  prior = prior,
  bf_thresholds = c(10, 1/10)
)
}
\arguments{
\item{df}{TODO}

\item{groups}{TODO}

\item{formula}{TODO}

\item{prior}{TODO}

\item{bf_thresholds}{TODO}
}
\value{
A tibble with the grouping variables, a calculated BF,
and the inference.
}
\description{
Run a Bayesian t-test.
}
