% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_bf_extractor.R
\name{my_bf_extractor}
\alias{my_bf_extractor}
\title{Function to exract BF values}
\usage{
my_bf_extractor(bf_obj)
}
\arguments{
\item{bf_obj}{the output object of a BayesFactor function}
}
\value{
A tibble containing the Bayes factor information.
}
\description{
This is the description
}
