% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_country.R
\name{prepare_plot_data_country}
\alias{prepare_plot_data_country}
\title{Prepare data for the cultural distance and effect size comparison}
\usage{
prepare_plot_data_country(data, study_name)
}
\arguments{
\item{data}{datatable to be plotted}

\item{study_name}{name of the study, either study1a, study1b, study2a, or study2b}
}
\value{
The function returns a datatable ready to be plotted.
}
\description{
This function ...
}
