% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data_study1.R
\name{prepare_plot_data_study1}
\alias{prepare_plot_data_study1}
\title{Prepare datatable for plotting for study1}
\usage{
prepare_plot_data_study1(data, study_type)
}
\arguments{
\item{data}{the datatable to be plotted}

\item{study_type}{the type of the study either "trolley" or "speedboat"}
}
\value{
the function returns a modified datatable
ready to be passed to the plotting function.
}
\description{
Prepares datatable for creating a plot for study1. The plot
can be created with the \code{\link{create_plot_study1}} function.
}
