% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_plot_data.R
\name{summarise_plot_data}
\alias{summarise_plot_data}
\title{Summarize data for plotting}
\usage{
summarise_plot_data(data)
}
\arguments{
\item{data}{the datatable containing the data to be plotted}
}
\value{
The function returns the summarized datatable.
}
\description{
The function summarizes data for plotting.
}
