% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_study.R
\name{summarise_study}
\alias{summarise_study}
\title{Function to summarize studies into variables}
\usage{
summarise_study(data = NULL)
}
\arguments{
\item{data}{datatable containing the raw values}
}
\value{
The function returns the summarized dataframe.
}
\description{
Creates a datatable of aggregated values based
on a datatable with the raw variables.
}
