% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_culture_utilitarianism.R
\name{tidy_culture_utilitarianism}
\alias{tidy_culture_utilitarianism}
\title{Tidy the results of the interaction analysis}
\usage{
tidy_culture_utilitarianism(data)
}
\arguments{
\item{data}{the result of the calculate_interaction_stats function}
}
\value{
tidy table
}
\description{
Lorem ipsum
}
