% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_interaction_stats.R
\name{tidy_interaction_stats}
\alias{tidy_interaction_stats}
\title{Tidy the results of the interaction analysis}
\usage{
tidy_interaction_stats(data)
}
\arguments{
\item{data}{the result of the calculate_interaction_stats function}
}
\value{
tidy table
}
\description{
Lorem ipsum
}
