% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_study1.R
\name{create_plot_study1}
\alias{create_plot_study1}
\title{Create a plot for study1}
\usage{
create_plot_study1(data, data_sum)
}
\arguments{
\item{data}{the datatable to be plotted}

\item{data_sum}{summarised plot data}
}
\value{
The function returns a ggplot2 list object.
}
\description{
The function creates a plot for the
trolley problem paper specifically from already prepared
data.
}
